Phase 6a: Extract rankings/alliances logic into useRankingsAlliances hook
Extract ~700 lines of rankings, alliances, and district rankings fetch
logic from App.jsx into a dedicated useRankingsAlliances hook, with an
EventStoreProvider serving as the context bridge.

Architecture:
- State ownership stays in App.jsx (rankings, rankingsOverride, alliances,
  allianceCount, districtRankings, playoffs) for React reactivity
- The hook receives state + setters via a deps object and owns only the
  fetch functions, helpers, and semantic actions
- EventStoreProvider calls the hook, exposes functions via storeRef so
  App.jsx callers (getSchedule, loadEvent) above the provider tree can
  invoke them, and adds getRanks/getAlliances/getDistrictRanks to the
  actions context

Functions extracted from App.jsx to useRankingsAlliances:
- getRanks() (~215 lines) — fetches from TBA/FRC API/Cheesy Arena/FTC
  sources with epoch guards for stale-response protection
- getAlliances() (~310 lines) — fetches alliance selections, builds
  lookup maps, applies playoff reserve edits
- getDistrictRanks() (~12 lines) — fetches district rankings
- fetchTBARankings() / fetchTBAAlliances() — TBA API helpers
- conformCheesyArenaRankings() / conformFTCOfflineRankings() /
  conformFTCOfflineAlliance() — data normalizers
- normalizeFtcGatoolAllianceRow() — FTC alliance row normalizer
- applyPlayoffReserveEdits / roundThreeOrReserveRoleLabel (moved imports)

Semantic actions exposed via storeRef:
- resetRankingsAlliancesState(preserveOfflineData) — resets all
  rankings/alliances state when switching events
- applyUserPrefs({ rankingsOverride, allianceCount }) — applies
  cloud-synced user preferences

EventStoreProvider features:
- Request deduplication infrastructure (createInflightTracker)
- Epoch-based stale-response guards (createEpochGuard)
- Ref-stabilized actions context (never changes identity)
- Explicit action surface enumeration (13 actions)

App.jsx integration:
- eventStoreRef bridges App.jsx functions to store-owned functions
- getSchedule calls eventStoreRef.current.getRanks/getAlliances
- loadEvent calls eventStoreRef.current.resetRankingsAlliancesState
- All rankings/alliances state flows through eventDataValue into
  EventDataContext for consumption by pages via useEventData()
- Removed 3 unused imports (applyPlayoffReserveEdits,
  roundThreeOrReserveRoleLabel, normalizeFtcGatoolAllianceRow)

Metrics:
- App.jsx: 6,467 -> 5,816 lines (-651 lines, -10.1%)
- New: useRankingsAlliances.js (735 lines)
- New: EventStoreProvider.jsx (146 lines)
- Build: compiles with 0 new warnings (4 pre-existing)
- Bundle: +1.21 kB gzipped (due to hook module overhead)
